added relationships to categories

+++ spec/features/simple_user_journey_spec.rb
@@ -60,6 +60,7 @@ feature "a simple linear user journey " do
       click_link 'New Work'
       expect(page).to have_content 'Submit New Work'
       expect(page).to have_button  'Select author(s)'
+      select @new_work.category.full_title, :from => 'work_category_id'
       fill_in 'Title', :with => @new_work.title
       select @new_work.year.year.to_s, :from => 'work_year_1i'
       select ("%02d" % @new_work.duration.hour), :from => 'work_duration_4i'
@@ -121,7 +122,10 @@ feature "a simple linear user journey " do
       # and we're done!
       #
     ensure
-      filenames.each { |f| File.unlink(f) }
+      #
+      # filenames might be blank if the spec has failed previously
+      #
+      filenames.each { |f| File.unlink(f) } unless filenames.blank?
     end
   end
